pyproject.toml
[tool.poetry]
name = "robonomics-wiki-deprecation-notifier"
version = "0.1.0"
description = "A tool to notify Robonomics Wiki contributors about possible article deprecations"
authors = ["arseniiarsenii <arseniivelichko2@gmail.com>"]
readme = "README.md"
packages = [{include = "robonomics_wiki_deprecation_notifier"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.22.2"
httpx = "^0.23.0"
loguru = "^0.6.0"
PyYAML = "^6.0"
yarl = "^1.8.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
